backend/pyproject.toml
[project]
name = "bourracho"
version = "0.1.0"
authors = [{ name = "Vladeptres", email = "feahmpo.92@gmail.com" }]

description = "Ezebarti"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "bcrypt>=4.3.0",
    "django>=3.2",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.4.3",
    "django-rest-swagger>=2.2.0",
    "emoji>=2.14.1",
    "loguru>=0.7.3",
    "mongomock>=4.3.0",
    "pydantic>=2.11.7",
    "pymongo>=4.13.2",
    "uvicorn>=0.35.0",
]

[tool.ruff]
# Apply ruff to all packages in the workspace
extend-exclude = [".venv", "build", "dist"]
line-length = 120
lint.select = ["E", "F", "I", "B"]
lint.ignore = []

[tool.setuptools]
include-package-data = true
packages.find.include = ['bourracho*', "bourracho/*.json"]
packages.find.exclude = ['bourracho.tests*']

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.3",
]
